+++ crates/haagenti-zstd/examples/precise_bitstream.rs
@@ -214,7 +214,7 @@ fn decode_and_trace(bitstream: &[u8], seq_count: usize) {
 
         // Calculate values
         let ll_value = get_ll_baseline(ll_code) + ll_extra as u32;
-        let offset_value = (1u32 << of_code) + of_extra as u32;
+        let offset_value = (1u32 << of_code) + of_extra;
         let ml_baseline = get_ml_baseline(ml_code);
         let ml_value = ml_baseline + ml_extra as u32;
